requirements.txt
# Core dependencies
anthropic>=0.39.0
openai>=1.12.0
ollama>=0.4.0

# Voice processing
pvporcupine>=3.0.0
pyaudio>=0.2.14
SpeechRecognition>=3.10.0
vosk>=0.3.45
pyttsx3>=2.90
gTTS>=2.5.0

# System control
psutil>=5.9.0
pyautogui>=0.9.54

# Web framework
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
python-multipart>=0.0.9
jinja2>=3.1.3

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
loguru>=0.7.2
aiofiles>=23.2.1
httpx>=0.27.0
rich>=13.7.0
click>=8.1.7
prompt-toolkit>=3.0.43

# Database
aiosqlite>=0.19.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
